Refactored code, put loading spinner in general utility components
