+++ frontend/src/pages/home/components/NewMeeting.tsx
@@ -7,7 +7,7 @@ import { useUserContext } from '../../../contexts/UserContext';
 import { AddMeetingRequest } from '@shared/requests/MeetingRequests';
 import DatePickerUtil from '../../../utility_components/DatePickerUtil';
 import { addOneHour, roundToHour } from '../../../utils/timeUtil';
-import LoadingSpinner from '../../../pages/roles/components/LoadingSpinner';
+import LoadingSpinner from '../../../utility_components/LoadingSpinner';
 
 
 const defaultValues = {
